refactor extract db logic to store
