--- article/http/request.go
@@ -1,82 +1,0 @@
-package http
-
-import (
-	"github.com/labstack/echo"
-	"github.com/gosimple/slug"
-	"github.com/xesina/golang-echo-realworld-example-app/article"
-)
-
-type articleCreateRequest struct {
-	Article struct {
-		Title       string   `json:"title" validate:"required"`
-		Description string   `json:"description" validate:"required"`
-		Body        string   `json:"body" validate:"required"`
-		Tags        []string `json:"tagList, omitempty"`
-	} `json:"article"`
-}
-
-func (r *articleCreateRequest) bind(c echo.Context, a *article.Article) error {
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return err
-	}
-	a.Title = r.Article.Title
-	a.Slug = slug.Make(r.Article.Title)
-	a.Description = r.Article.Description
-	a.Body = r.Article.Body
-	if r.Article.Tags != nil {
-		for _, t := range r.Article.Tags {
-			a.Tags = append(a.Tags, article.Tag{Tag: t})
-		}
-	}
-	return nil
-}
-
-type articleUpdateRequest struct {
-	Article struct {
-		Title       string   `json:"title"`
-		Description string   `json:"description"`
-		Body        string   `json:"body"`
-		Tags        []string `json:"tagList"`
-	} `json:"article"`
-}
-
-func (r *articleUpdateRequest) populate(a *article.Article) {
-	r.Article.Title = a.Title
-	r.Article.Description = a.Description
-	r.Article.Body = a.Body
-}
-
-func (r *articleUpdateRequest) bind(c echo.Context, a *article.Article) error {
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return err
-	}
-	a.Title = r.Article.Title
-	a.Slug = slug.Make(a.Title)
-	a.Description = r.Article.Description
-	a.Body = r.Article.Body
-	return nil
-}
-
-type createCommentRequest struct {
-	Comment struct {
-		Body string `json:"body" validate:"required"`
-	} `json:"comment"`
-}
-
-func (r *createCommentRequest) bind(c echo.Context, cm *article.Comment) error {
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return err
-	}
-	cm.Body = r.Comment.Body
-	cm.UserID = userIDFromToken(c)
-	return nil
-}
