+++ handler/request.go
@@ -1,33 +1,10 @@
 package handler
 
 import (
-	"time"
-
-	"github.com/gosimple/slug"
 	"github.com/labstack/echo"
-	"github.com/xesina/golang-echo-realworld-example-app/models"
-	"github.com/xesina/golang-echo-realworld-example-app/utils"
+	"github.com/gosimple/slug"
+	"github.com/xesina/golang-echo-realworld-example-app/model"
 )
-
-type userResponse struct {
-	User struct {
-		Username string  `json:"username"`
-		Email    string  `json:"email"`
-		Bio      *string `json:"bio"`
-		Image    *string `json:"image"`
-		Token    string  `json:"token"`
-	} `json:"user"`
-}
-
-func newUserResponse(u *models.User) *userResponse {
-	r := new(userResponse)
-	r.User.Username = u.Username
-	r.User.Email = u.Email
-	r.User.Bio = u.Bio
-	r.User.Image = u.Image
-	r.User.Token = utils.GenerateJWT(u.ID)
-	return r
-}
 
 type userUpdateRequest struct {
 	User struct {
@@ -43,7 +20,7 @@ func newUserUpdateRequest() *userUpdateRequest {
 	return new(userUpdateRequest)
 }
 
-func (r *userUpdateRequest) populate(u *models.User) {
+func (r *userUpdateRequest) populate(u *model.User) {
 	r.User.Username = u.Username
 	r.User.Email = u.Email
 	r.User.Password = u.Password
@@ -55,7 +32,7 @@ func (r *userUpdateRequest) populate(u *models.User) {
 	}
 }
 
-func (r *userUpdateRequest) bind(c echo.Context, u *models.User) error {
+func (r *userUpdateRequest) bind(c echo.Context, u *model.User) error {
 	if err := c.Bind(r); err != nil {
 		return err
 	}
@@ -65,9 +42,11 @@ func (r *userUpdateRequest) bind(c echo.Context, u *models.User) error {
 	u.Username = r.User.Username
 	u.Email = r.User.Email
 	if r.User.Password != u.Password {
-		if err := u.HashPassword(r.User.Password); err != nil {
+		h, err := u.HashPassword(r.User.Password)
+		if err != nil {
 			return err
 		}
+		u.Password = h
 	}
 	u.Bio = &r.User.Bio
 	u.Image = &r.User.Image
@@ -82,7 +61,7 @@ type userRegisterRequest struct {
 	} `json:"user"`
 }
 
-func (r *userRegisterRequest) bind(c echo.Context, u *models.User) error {
+func (r *userRegisterRequest) bind(c echo.Context, u *model.User) error {
 	if err := c.Bind(r); err != nil {
 		return err
 	}
@@ -91,10 +70,11 @@ func (r *userRegisterRequest) bind(c echo.Context, u *models.User) error {
 	}
 	u.Username = r.User.Username
 	u.Email = r.User.Email
-	if err := u.HashPassword(r.User.Password); err != nil {
+	h, err := u.HashPassword(r.User.Password)
+	if err != nil {
 		return err
 	}
-
+	u.Password = h
 	return nil
 }
 
@@ -105,120 +85,14 @@ type userLoginRequest struct {
 	} `json:"user"`
 }
 
-func (r *userLoginRequest) bind(c echo.Context, u *models.User) error {
+func (r *userLoginRequest) bind(c echo.Context) error {
 	if err := c.Bind(r); err != nil {
 		return err
 	}
 	if err := c.Validate(r); err != nil {
 		return err
 	}
-	u.Email = r.User.Email
-	if err := u.HashPassword(r.User.Password); err != nil {
-		return err
-	}
-
 	return nil
-}
-
-type profileResponse struct {
-	Profile struct {
-		Username  string  `json:"username"`
-		Bio       *string `json:"bio"`
-		Image     *string `json:"image"`
-		Following bool    `json:"following"`
-	} `json:"profile"`
-}
-
-func newProfileResponse(c echo.Context, u *models.User) *profileResponse {
-	r := new(profileResponse)
-	r.Profile.Username = u.Username
-	r.Profile.Bio = u.Bio
-	r.Profile.Image = u.Image
-	r.Profile.Following = u.FollowedBy(userIDFromToken(c))
-	return r
-}
-
-type articleResponse struct {
-	Slug           string    `json:"slug"`
-	Title          string    `json:"title"`
-	Description    string    `json:"description"`
-	Body           string    `json:"body"`
-	TagList        []string  `json:"tagList"`
-	CreatedAt      time.Time `json:"createdAt"`
-	UpdatedAt      time.Time `json:"updatedAt"`
-	Favorited      bool      `json:"favorited"`
-	FavoritesCount int       `json:"favoritesCount"`
-	Author         struct {
-		Username  string  `json:"username"`
-		Bio       *string `json:"bio"`
-		Image     *string `json:"image"`
-		Following bool    `json:"following"`
-	} `json:"author"`
-}
-
-type singleArticleResponse struct {
-	Article *articleResponse `json:"article"`
-}
-
-type articleListResponse struct {
-	Articles      []*articleResponse `json:"articles"`
-	ArticlesCount int                `json:"articlesCount"`
-}
-
-func newArticleResponse(c echo.Context, a *models.Article) *singleArticleResponse {
-	ar := new(articleResponse)
-	ar.TagList = make([]string, 0)
-	ar.Slug = a.Slug
-	ar.Title = a.Title
-	ar.Description = a.Description
-	ar.Body = a.Body
-	ar.CreatedAt = a.CreatedAt
-	ar.UpdatedAt = a.UpdatedAt
-	for _, t := range a.Tags {
-		ar.TagList = append(ar.TagList, t.Tag)
-	}
-	for _, u := range a.Favorites {
-		if u.ID == userIDFromToken(c) {
-			ar.Favorited = true
-		}
-	}
-	ar.FavoritesCount = len(a.Favorites)
-	ar.Author.Username = a.Author.Username
-	ar.Author.Image = a.Author.Image
-	ar.Author.Bio = a.Author.Bio
-	ar.Author.Following = a.Author.FollowedBy(userIDFromToken(c))
-	return &singleArticleResponse{ar}
-}
-
-func newArticleListResponse(c echo.Context, articles []models.Article, count int) *articleListResponse {
-	r := new(articleListResponse)
-	r.Articles = make([]*articleResponse, 0)
-	for _, a := range articles {
-		ar := new(articleResponse)
-		ar.TagList = make([]string, 0)
-		ar.Slug = a.Slug
-		ar.Title = a.Title
-		ar.Description = a.Description
-		ar.Body = a.Body
-		ar.CreatedAt = a.CreatedAt
-		ar.UpdatedAt = a.UpdatedAt
-		for _, t := range a.Tags {
-			ar.TagList = append(ar.TagList, t.Tag)
-		}
-		for _, u := range a.Favorites {
-			if u.ID == userIDFromToken(c) {
-				ar.Favorited = true
-			}
-		}
-		ar.FavoritesCount = len(a.Favorites)
-		ar.Author.Username = a.Author.Username
-		ar.Author.Image = a.Author.Image
-		ar.Author.Bio = a.Author.Bio
-		ar.Author.Following = a.Author.FollowedBy(userIDFromToken(c))
-		r.Articles = append(r.Articles, ar)
-	}
-	r.ArticlesCount = count
-	return r
 }
 
 type articleCreateRequest struct {
@@ -230,7 +104,7 @@ type articleCreateRequest struct {
 	} `json:"article"`
 }
 
-func (r *articleCreateRequest) bind(c echo.Context, a *models.Article) error {
+func (r *articleCreateRequest) bind(c echo.Context, a *model.Article) error {
 	if err := c.Bind(r); err != nil {
 		return err
 	}
@@ -243,7 +117,7 @@ func (r *articleCreateRequest) bind(c echo.Context, a *models.Article) error {
 	a.Body = r.Article.Body
 	if r.Article.Tags != nil {
 		for _, t := range r.Article.Tags {
-			a.Tags = append(a.Tags, models.Tag{Tag: t})
+			a.Tags = append(a.Tags, model.Tag{Tag: t})
 		}
 	}
 	return nil
@@ -258,13 +132,13 @@ type articleUpdateRequest struct {
 	} `json:"article"`
 }
 
-func (r *articleUpdateRequest) populate(a *models.Article) {
+func (r *articleUpdateRequest) populate(a *model.Article) {
 	r.Article.Title = a.Title
 	r.Article.Description = a.Description
 	r.Article.Body = a.Body
 }
 
-func (r *articleUpdateRequest) bind(c echo.Context, a *models.Article) error {
+func (r *articleUpdateRequest) bind(c echo.Context, a *model.Article) error {
 	if err := c.Bind(r); err != nil {
 		return err
 	}
@@ -284,7 +158,7 @@ type createCommentRequest struct {
 	} `json:"comment"`
 }
 
-func (r *createCommentRequest) bind(c echo.Context, cm *models.Comment) error {
+func (r *createCommentRequest) bind(c echo.Context, cm *model.Comment) error {
 	if err := c.Bind(r); err != nil {
 		return err
 	}
@@ -294,69 +168,4 @@ func (r *createCommentRequest) bind(c echo.Context, cm *models.Comment) error {
 	cm.Body = r.Comment.Body
 	cm.UserID = userIDFromToken(c)
 	return nil
-}
-
-type commentResponse struct {
-	ID        uint      `json:"id"`
-	Body      string    `json:"body"`
-	CreatedAt time.Time `json:"createdAt"`
-	UpdatedAt time.Time `json:"updatedAt"`
-	Author    struct {
-		Username  string  `json:"username"`
-		Bio       *string `json:"bio"`
-		Image     *string `json:"image"`
-		Following bool    `json:"following"`
-	} `json:"author"`
-}
-
-type singleCommentResponse struct {
-	Comment *commentResponse `json:"comment"`
-}
-
-type commentListResponse struct {
-	Comments []commentResponse `json:"comments"`
-}
-
-func newCommentResponse(c echo.Context, cm *models.Comment) *singleCommentResponse {
-	comment := new(commentResponse)
-	comment.ID = cm.ID
-	comment.Body = cm.Body
-	comment.CreatedAt = cm.CreatedAt
-	comment.UpdatedAt = cm.UpdatedAt
-	comment.Author.Username = cm.User.Username
-	comment.Author.Image = cm.User.Image
-	comment.Author.Bio = cm.User.Bio
-	comment.Author.Following = cm.User.FollowedBy(userIDFromToken(c))
-	return &singleCommentResponse{comment}
-}
-
-func newCommentListResponse(c echo.Context, comments []models.Comment) *commentListResponse {
-	r := new(commentListResponse)
-	cr := commentResponse{}
-	r.Comments = make([]commentResponse, 0)
-	for _, i := range comments {
-		cr.ID = i.ID
-		cr.Body = i.Body
-		cr.CreatedAt = i.CreatedAt
-		cr.UpdatedAt = i.UpdatedAt
-		cr.Author.Username = i.User.Username
-		cr.Author.Image = i.User.Image
-		cr.Author.Bio = i.User.Bio
-		cr.Author.Following = i.User.FollowedBy(userIDFromToken(c))
-
-		r.Comments = append(r.Comments, cr)
-	}
-	return r
-}
-
-type tagListResponse struct {
-	Tags []string `json:"tags"`
-}
-
-func newTagListResponse(tags []models.Tag) *tagListResponse {
-	r := new(tagListResponse)
-	for _, t := range tags {
-		r.Tags = append(r.Tags, t.Tag)
-	}
-	return r
 }
