+++ handler/response.go
@@ -1,10 +1,50 @@
-package http
+package handler
 
 import (
 	"time"
 	"github.com/labstack/echo"
-	"github.com/xesina/golang-echo-realworld-example-app/article"
+	"github.com/xesina/golang-echo-realworld-example-app/model"
+	"github.com/xesina/golang-echo-realworld-example-app/utils"
+	"github.com/xesina/golang-echo-realworld-example-app/user"
 )
+
+type userResponse struct {
+	User struct {
+		Username string  `json:"username"`
+		Email    string  `json:"email"`
+		Bio      *string `json:"bio"`
+		Image    *string `json:"image"`
+		Token    string  `json:"token"`
+	} `json:"user"`
+}
+
+func newUserResponse(u *model.User) *userResponse {
+	r := new(userResponse)
+	r.User.Username = u.Username
+	r.User.Email = u.Email
+	r.User.Bio = u.Bio
+	r.User.Image = u.Image
+	r.User.Token = utils.GenerateJWT(u.ID)
+	return r
+}
+
+type profileResponse struct {
+	Profile struct {
+		Username  string  `json:"username"`
+		Bio       *string `json:"bio"`
+		Image     *string `json:"image"`
+		Following bool    `json:"following"`
+	} `json:"profile"`
+}
+
+func newProfileResponse(us user.Store, userID uint, u *model.User) *profileResponse {
+	r := new(profileResponse)
+	r.Profile.Username = u.Username
+	r.Profile.Bio = u.Bio
+	r.Profile.Image = u.Image
+	r.Profile.Following, _ = us.IsFollower(u.ID, userID)
+	return r
+}
 
 type articleResponse struct {
 	Slug           string    `json:"slug"`
@@ -33,7 +73,7 @@ type articleListResponse struct {
 	ArticlesCount int                `json:"articlesCount"`
 }
 
-func newArticleResponse(c echo.Context, a *article.Article) *singleArticleResponse {
+func newArticleResponse(c echo.Context, a *model.Article) *singleArticleResponse {
 	ar := new(articleResponse)
 	ar.TagList = make([]string, 0)
 	ar.Slug = a.Slug
@@ -58,7 +98,7 @@ func newArticleResponse(c echo.Context, a *article.Article) *singleArticleRespon
 	return &singleArticleResponse{ar}
 }
 
-func newArticleListResponse(c echo.Context, articles []article.Article, count int) *articleListResponse {
+func newArticleListResponse(us user.Store, userID uint, articles []model.Article, count int) *articleListResponse {
 	r := new(articleListResponse)
 	r.Articles = make([]*articleResponse, 0)
 	for _, a := range articles {
@@ -74,7 +114,7 @@ func newArticleListResponse(c echo.Context, articles []article.Article, count in
 			ar.TagList = append(ar.TagList, t.Tag)
 		}
 		for _, u := range a.Favorites {
-			if u.ID == userIDFromToken(c) {
+			if u.ID == userID {
 				ar.Favorited = true
 			}
 		}
@@ -82,7 +122,7 @@ func newArticleListResponse(c echo.Context, articles []article.Article, count in
 		ar.Author.Username = a.Author.Username
 		ar.Author.Image = a.Author.Image
 		ar.Author.Bio = a.Author.Bio
-		ar.Author.Following = a.Author.FollowedBy(userIDFromToken(c))
+		ar.Author.Following, _ = us.IsFollower(a.AuthorID, userID)
 		r.Articles = append(r.Articles, ar)
 	}
 	r.ArticlesCount = count
@@ -110,7 +150,7 @@ type commentListResponse struct {
 	Comments []commentResponse `json:"comments"`
 }
 
-func newCommentResponse(c echo.Context, cm *article.Comment) *singleCommentResponse {
+func newCommentResponse(c echo.Context, cm *model.Comment) *singleCommentResponse {
 	comment := new(commentResponse)
 	comment.ID = cm.ID
 	comment.Body = cm.Body
@@ -123,7 +163,7 @@ func newCommentResponse(c echo.Context, cm *article.Comment) *singleCommentRespo
 	return &singleCommentResponse{comment}
 }
 
-func newCommentListResponse(c echo.Context, comments []article.Comment) *commentListResponse {
+func newCommentListResponse(c echo.Context, comments []model.Comment) *commentListResponse {
 	r := new(commentListResponse)
 	cr := commentResponse{}
 	r.Comments = make([]commentResponse, 0)
@@ -146,7 +186,7 @@ type tagListResponse struct {
 	Tags []string `json:"tags"`
 }
 
-func newTagListResponse(tags []article.Tag) *tagListResponse {
+func newTagListResponse(tags []model.Tag) *tagListResponse {
 	r := new(tagListResponse)
 	for _, t := range tags {
 		r.Tags = append(r.Tags, t.Tag)
