+++ model/article.go
@@ -1,0 +1,33 @@
+package model
+
+import (
+	"github.com/jinzhu/gorm"
+)
+
+type Article struct {
+	gorm.Model
+	Slug        string `gorm:"unique_index;not null"`
+	Title       string `gorm:"not null"`
+	Description string
+	Body        string
+	Author      User
+	AuthorID    uint
+	Comments    []Comment
+	Favorites   []User `gorm:"many2many:favorites;"`
+	Tags        []Tag  `gorm:"many2many:article_tags;association_autocreate:false"`
+}
+
+type Comment struct {
+	gorm.Model
+	Article   Article
+	ArticleID uint
+	User      User
+	UserID    uint
+	Body      string
+}
+
+type Tag struct {
+	gorm.Model
+	Tag      string    `gorm:"unique_index"`
+	Articles []Article `gorm:"many2many:article_tags;"`
+}
