+++ store/user.go
@@ -1,0 +1,86 @@
+package store
+
+import (
+	"github.com/jinzhu/gorm"
+	"github.com/xesina/golang-echo-realworld-example-app/model"
+)
+
+type UserStore struct {
+	db *gorm.DB
+}
+
+func NewUserStore(db *gorm.DB) *UserStore {
+	return &UserStore{
+		db: db,
+	}
+}
+
+func (us *UserStore) GetByID(id uint) (*model.User, error) {
+	var m model.User
+	if err := us.db.First(&m, id).Error; err != nil {
+		if gorm.IsRecordNotFoundError(err) {
+			return nil, nil
+		}
+		return nil, err
+	}
+	return &m, nil
+}
+
+func (us *UserStore) GetByEmail(e string) (*model.User, error) {
+	var m model.User
+	if err := us.db.Where(&model.User{Email: e}).First(&m).Error; err != nil {
+		if gorm.IsRecordNotFoundError(err) {
+			return nil, nil
+		}
+		return nil, err
+	}
+	return &m, nil
+}
+
+func (us *UserStore) GetByUsername(username string) (*model.User, error) {
+	var m model.User
+	if err := us.db.Where(&model.User{Username: username}).Preload("Followers").First(&m).Error; err != nil {
+		if gorm.IsRecordNotFoundError(err) {
+			return nil, nil
+		}
+		return nil, err
+	}
+	return &m, nil
+}
+
+func (us *UserStore) Create(u *model.User) (err error) {
+	return us.db.Create(u).Error
+}
+
+func (us *UserStore) Update(u *model.User) error {
+	return us.db.Model(u).Update(u).Error
+}
+
+func (us *UserStore) AddFollower(u *model.User, followerID uint) error {
+	return us.db.Model(u).Association("Followers").Append(&model.Follow{FollowerID: followerID, FollowingID: u.ID}).Error
+}
+
+func (us *UserStore) RemoveFollower(u *model.User, followerID uint) error {
+	f := model.Follow{
+		FollowerID:  followerID,
+		FollowingID: u.ID,
+	}
+	if err := us.db.Model(u).Association("Followers").Find(&f).Error; err != nil {
+		return err
+	}
+	if err := us.db.Delete(f).Error; err != nil {
+		return err
+	}
+	return nil
+}
+
+func (us *UserStore) IsFollower(userID, followerID uint) (bool, error) {
+	var f model.Follow
+	if err := us.db.Where("following_id = ? AND follower_id = ?", userID, followerID).Find(&f).Error; err != nil {
+		if gorm.IsRecordNotFoundError(err) {
+			return false, nil
+		}
+		return false, err
+	}
+	return true, nil
+}
