--- user/http/request.go
@@ -1,95 +1,0 @@
-package http
-
-import (
-	"github.com/labstack/echo"
-	"github.com/xesina/golang-echo-realworld-example-app/user"
-)
-
-type userUpdateRequest struct {
-	User struct {
-		Username string `json:"username"`
-		Email    string `json:"email" validate:"email"`
-		Password string `json:"password"`
-		Bio      string `json:"bio"`
-		Image    string `json:"image"`
-	} `json:"user"`
-}
-
-func newUserUpdateRequest() *userUpdateRequest {
-	return new(userUpdateRequest)
-}
-
-func (r *userUpdateRequest) populate(u *user.User) {
-	r.User.Username = u.Username
-	r.User.Email = u.Email
-	r.User.Password = u.Password
-	if u.Bio != nil {
-		r.User.Bio = *u.Bio
-	}
-	if u.Image != nil {
-		r.User.Image = *u.Image
-	}
-}
-
-func (r *userUpdateRequest) bind(c echo.Context, u *user.User) error {
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return err
-	}
-	u.Username = r.User.Username
-	u.Email = r.User.Email
-	if r.User.Password != u.Password {
-		h, err := u.HashPassword(r.User.Password)
-		if err != nil {
-			return err
-		}
-		u.Password = h
-	}
-	u.Bio = &r.User.Bio
-	u.Image = &r.User.Image
-	return nil
-}
-
-type userRegisterRequest struct {
-	User struct {
-		Username string `json:"username" validate:"required"`
-		Email    string `json:"email" validate:"required,email"`
-		Password string `json:"password" validate:"required"`
-	} `json:"user"`
-}
-
-func (r *userRegisterRequest) bind(c echo.Context, u *user.User) error {
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return err
-	}
-	u.Username = r.User.Username
-	u.Email = r.User.Email
-	h, err := u.HashPassword(r.User.Password)
-	if err != nil {
-		return err
-	}
-	u.Password = h
-	return nil
-}
-
-type userLoginRequest struct {
-	User struct {
-		Email    string `json:"email" validate:"required,email"`
-		Password string `json:"password" validate:"required"`
-	} `json:"user"`
-}
-
-func (r *userLoginRequest) bind(c echo.Context) error {
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return err
-	}
-	return nil
-}
