--- user/http/response.go
@@ -1,45 +1,0 @@
-package http
-
-import (
-	"github.com/labstack/echo"
-	"github.com/xesina/golang-echo-realworld-example-app/user"
-	"github.com/xesina/golang-echo-realworld-example-app/utils"
-)
-
-type userResponse struct {
-	User struct {
-		Username string  `json:"username"`
-		Email    string  `json:"email"`
-		Bio      *string `json:"bio"`
-		Image    *string `json:"image"`
-		Token    string  `json:"token"`
-	} `json:"user"`
-}
-
-func newUserResponse(u *user.User) *userResponse {
-	r := new(userResponse)
-	r.User.Username = u.Username
-	r.User.Email = u.Email
-	r.User.Bio = u.Bio
-	r.User.Image = u.Image
-	r.User.Token = utils.GenerateJWT(u.ID)
-	return r
-}
-
-type profileResponse struct {
-	Profile struct {
-		Username  string  `json:"username"`
-		Bio       *string `json:"bio"`
-		Image     *string `json:"image"`
-		Following bool    `json:"following"`
-	} `json:"profile"`
-}
-
-func newProfileResponse(c echo.Context, u *user.User) *profileResponse {
-	r := new(profileResponse)
-	r.Profile.Username = u.Username
-	r.Profile.Bio = u.Bio
-	r.Profile.Image = u.Image
-	r.Profile.Following = u.FollowedBy(userIDFromToken(c))
-	return r
-}
