+++ utils/errors.go
@@ -38,3 +38,10 @@ func AccessForbidden() Error {
 	e.Errors["body"] = "access forbidden"
 	return e
 }
+
+func NotFound() Error {
+	e := Error{}
+	e.Errors = make(map[string]interface{})
+	e.Errors["body"] = "resource not found"
+	return e
+}
